Update docs to reflect CHUNK_TYPE configuration and TS.INFO's debug flag. Also fix `bytesPerSample` to respond with a float instead of an integer.

+++ src/module.c
@@ -384,7 +384,7 @@ int TSDB_info(RedisModuleCtx *ctx, RedisModuleString **argv, int argc) {
             RedisModule_ReplyWithSimpleString(ctx, "size");
             RedisModule_ReplyWithLongLong(ctx, chunkSize);
             RedisModule_ReplyWithSimpleString(ctx, "bytesPerSample");
-            RedisModule_ReplyWithDouble(ctx, chunkSize / numOfSamples);
+            RedisModule_ReplyWithDouble(ctx, (float)chunkSize / numOfSamples);
             chunkCount++;
         }
         RedisModule_DictIteratorStop(iter);
